tidied tests
