+++ cli/package.json
@@ -70,7 +70,6 @@
     "@types/lodash.compact": "^3.0.7",
     "@types/lodash.get": "^4.4.7",
     "@types/lodash.set": "^4.3.7",
-    "@types/mock-fs": "^4.13.1",
     "@types/relaxed-json": "^1.0.1",
     "@types/text-table": "^0.2.2",
     "@types/tmp": "^0.2.3",
@@ -78,7 +77,6 @@
     "eslint": "^8.41.0",
     "eslint-config-oclif": "^4.0.0",
     "eslint-config-oclif-typescript": "^1.0.3",
-    "mock-fs": "^5.2.0",
     "oclif": "^3.9.0",
     "shx": "^0.3.4",
     "ts-node": "^10.9.1",
